requirements.txt
azure-cognitiveservices-speech
elevenlabs
pygame
python-dotenv
